requirements.txt
numpy
scikit-learn 
matplotlib 
pandas
torchinfo 
tqdm
pyyaml
regex
requests